Move datepicker stuff into date folder
