+++ backoffice/src/components/common/page/form/date/Date.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables';
+@import '../../../../../styles/variables';
 
 .Date {
