+++ backoffice/src/index.js
@@ -5,7 +5,6 @@ import ReactDOM from 'react-dom';
 import 'popper.js/dist/popper.min.js';
 import 'bootstrap/dist/js/bootstrap.min.js';
 import '@fortawesome/fontawesome-free/css/all.min.css';
-import './utils/datepicker.js';
 import './utils/moment-fr';
 import App from './App';
 import * as Sentry from './utils/sentry';
